requirements.txt
selenium=4.4.0
Pillow=9.2.0
pytesseract=0.3.10
python-decouple